bugs in hellobert (argument type recompute_huffman); make bert deal with start_address!=0 ... and within a repeat; bert deal with updated compress we_bits at BRAM18 level

+++ docs/tutorials/huffman/timing/hellobert.c
@@ -172,7 +172,7 @@ void compare(int hlen, int rlen)
     }
 }
 
-void recompute_huffman(int *result_code) {
+void recompute_huffman(uint64_t *result_code) {
   UChar len[256];
   Int32 code[256];
